actions/requirements.txt
pymysql
mysql-connector-python
# Add more dependencies here



